fixes #168 Fake texts Minigame System
- added fake history blocks
- player needs to drag all real history blocks and leave fake blocks in order to win
