+++ RaymondMaarloeve/Assets/Scenes/Game.unity
@@ -3433,7 +3433,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: a7a77b12dd6bc29f8bebb7f7f4e17154, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  npcCount: 6
   npcPrefabs:
   - {fileID: 1668834221483606419, guid: 363a0160b72821de083880c20963240f, type: 3}
   - {fileID: 2966672667916528032, guid: 050a8da0f92dbd6eb8233cf0fbce4f96, type: 3}
@@ -3608,13 +3607,6 @@ MonoBehaviour:
   cancelButton: {fileID: 1435796869}
   resultText: {fileID: 1821656902}
   murderDescription: 
-  blockTexts:
-  - Culprit lost his job and went bankrupt.
-  - While attending the mass he saw the golden grail.
-  - Culprit bribed the priest to get the keys to the church.
-  - He waited until midnight to break into the church and steal the grail.
-  - While attempting the theft he was seen by the guard.
-  - Culprit killed the guard and got away.
   triggerDay: 3
 --- !u!1 &1310063284
 GameObject:
