--- RaymondMaarloeve/Assets/Scripts/HistoryBlockDrag.cs
@@ -1,43 +1,0 @@
-﻿using UnityEngine;
-using UnityEngine.EventSystems;
-using TMPro;
-
-[RequireComponent(typeof(CanvasGroup))]
-public class HistoryBlockDrag : MonoBehaviour, IBeginDragHandler, IDragHandler, IEndDragHandler
-{
-    private CanvasGroup canvasGroup;
-    private RectTransform rect;
-    private Transform originalParent;
-    private Vector2 originalPosition;
-    private Canvas rootCanvas;
-
-    void Awake()
-    {
-        canvasGroup = GetComponent<CanvasGroup>();
-        rect = GetComponent<RectTransform>();
-        rootCanvas = GetComponentInParent<Canvas>();
-    }
-
-    public void OnBeginDrag(PointerEventData eventData)
-    {
-        originalParent = transform.parent;
-        originalPosition = rect.anchoredPosition;
-        transform.SetParent(rootCanvas.transform);
-        canvasGroup.blocksRaycasts = false;
-    }
-
-    public void OnDrag(PointerEventData eventData)
-    {
-        rect.anchoredPosition += eventData.delta / rootCanvas.scaleFactor;
-    }
-
-    public void OnEndDrag(PointerEventData eventData)
-    {
-        if (transform.parent == rootCanvas.transform)
-        {
-            transform.SetParent(originalParent);
-            rect.anchoredPosition = originalPosition;
-        }
-        canvasGroup.blocksRaycasts = true;
-    }
-}
