+++ RaymondMaarloeve/Assets/Scripts/MiniGame.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 80cf2bb70b7cd0a42a34fc6850bdf7a3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
